Done. Going to go through it really quick when i get home and then deploy to heroku
