+++ keap-signup-frontend/src/components/FormArea/ErrorSignUp.js
@@ -1,0 +1,13 @@
+const ErrorSignUp = () => {
+  return (
+    <div className="container flex md:w-[220px] md:ml-3">
+      <p className="font-open-sans text-Red mt-3 md:mt-2 sm:text-mobile-errors">
+        There was a problem submitting.
+        <br />
+        Please try again.
+      </p>
+    </div>
+  );
+};
+
+export default ErrorSignUp;
