+++ keap-signup-frontend/src/components/FormArea/SuccessSignUp.js
@@ -1,0 +1,14 @@
+const SuccessSignUp = () => {
+  return (
+    <div className="flex flex-col w-[96%] h-[414px] items-center justify-center">
+      <h2 className="font-roboto sm:text-mobile-h2 md:text-desktop-h2">
+        🎉 Thank you! 🎉
+      </h2>
+      <p className="container flex justify-center text-center font-open-sans w-[320px] md:w-[350px] md:mt-2 sm:text-mobile-body">
+        You're well on your way to automating your business!
+      </p>
+    </div>
+  );
+};
+
+export default SuccessSignUp;
