+++ keap-signup-frontend/src/components/FormArea/TermsAndService.js
@@ -1,0 +1,10 @@
+const TermsAndService = () => {
+  return (
+    <p className="container flex text-Grey mt-3 mb-4 md:mt-5 sm:text-mobile-legal">
+      By clicking submit you agree to our
+      <mark className="text-Green bg-transparent">&nbsp;Terms and Service</mark>
+    </p>
+  );
+};
+
+export default TermsAndService;
